Added Hylics and OFF enemies.

+++ src/main/java/net/tyconxon/offmod/procedures/ATMwithdrawProcedure.java
@@ -135,7 +135,7 @@ public class ATMwithdrawProcedure {
 				}
 			}
 			if (world instanceof World && !world.isRemote()) {
-				((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+				((World) world).playSound(null, new BlockPos(x, y, z),
 						(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("offmod:cash_register")),
 						SoundCategory.NEUTRAL, (float) 1, (float) 1);
 			} else {
@@ -244,7 +244,7 @@ public class ATMwithdrawProcedure {
 				}
 			}
 			if (world instanceof World && !world.isRemote()) {
-				((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+				((World) world).playSound(null, new BlockPos(x, y, z),
 						(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("offmod:cash_register")),
 						SoundCategory.NEUTRAL, (float) 1, (float) 1);
 			} else {
@@ -254,7 +254,7 @@ public class ATMwithdrawProcedure {
 			}
 		} else {
 			if (world instanceof World && !world.isRemote()) {
-				((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+				((World) world).playSound(null, new BlockPos(x, y, z),
 						(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("offmod:error")),
 						SoundCategory.NEUTRAL, (float) 1, (float) 1);
 			} else {
