+++ src/main/java/net/tyconxon/offmod/procedures/AbaddonsMeatFoodEatenProcedure.java
@@ -61,7 +61,7 @@ public class AbaddonsMeatFoodEatenProcedure {
 		if (entity instanceof LivingEntity)
 			((LivingEntity) entity).addPotionEffect(new EffectInstance(Effects.WEAKNESS, (int) 100, (int) 20));
 		if (world instanceof World && !world.isRemote()) {
-			((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+			((World) world).playSound(null, new BlockPos(x, y, z),
 					(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("offmod:itemuse")),
 					SoundCategory.PLAYERS, (float) 1, (float) 1);
 		} else {
