+++ src/main/java/net/tyconxon/offmod/procedures/YourFatherBlockValidPlacementConditionProcedure.java
@@ -34,7 +34,7 @@ public class YourFatherBlockValidPlacementConditionProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (world.isAirBlock(new BlockPos((int) x, (int) (y - 1), (int) z))) {
+		if (world.isAirBlock(new BlockPos(x, y - 1, z))) {
 			return false;
 		}
 		return true;
